// app/types/Characters.ts
export interface Characters {
    id: number;
    name: string;
    status: string;
    species: string;
    type: string;
    gender: string;
    origin: {
        name: string;   
        url: string;
    };
    location: {
        name: string;   
        url: string;
    };
    image: string;
    episode: string[];
    url: string;
    created: string;
}

export interface Info {
    count: number;
    pages: number;
    next: string | null;
    prev: string | null;
}

export interface CharactersResponse {
    info: Info;
    results: Characters[];
}